Longer Level, more enemies and more scenarie
I made the level longer so that it would not be over so soon and with the longer level I added in 3 more enemires making two of them copies of the original and the last one a bigger Boss type enemy the player will face in the cave. I also updated the sound the player makes when meleeing and added in more scenarie so the landscape the player walks through doesnt seem so barren.